# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  models.LoginResponseDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      token:
        type: string
      user_id:
        type: integer
    type: object
  models.LoginUserDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.RegisterUserDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.Role:
    enum:
    - user
    - moderator
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
...
  models.UpdateUserDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API для управления событиями, публикациями и изображениями.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: RIP API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Осуществляет вход пользователя, если его учетные данные верны,
        и возвращает JWT токен.
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponseDTO'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Вход пользователя
      tags:
      - Auth
  /users/logout:
    post:
      consumes:
      - application/json
      description: Завершаем сессию пользователя и удаляем его токен.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Неавторизованный пользователь
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Выход пользователя
      tags:
      - Auth
  /users/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя и возвращает JWT токен.
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LoginResponseDTO'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /users/update:
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный пользователь
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неавторизованный пользователь
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить информацию о пользователе
      tags:
      - Auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Use the JWT Bearer Token for authorization. Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"